testingroles

+++ app/Http/Controllers/ManageuserController.php
@@ -5,23 +5,42 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Password;
 use App\Models\User;
+use App\Models\Role;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Contracts\Auth\Access\Gate;
 
 class ManageuserController extends Controller
 {
 
+    private int $superadmin;
+
+    public function __construct()
+    {
+        $this->superadmin = Role::superadmin();
+        // $this->admin = Role::admin();
+    }
+
     public function index()
     {
+        // dd(auth()->user()->role->hasAuthority());
+        // dd(User::find(1)->role);
+        // $dat = Role::higherAuthority(3, 2);
+        // dd($dat);
+        // dd(auth()->user()->role_id);
+
         $this->authorize('manage-accounts');
-        return view('auth.pages.Users.manageuser', ['users' => User::all()]);
+        if (auth()->user()->role_id == $this->superadmin)
+            return view('auth.pages.Users.manageuser', ['users' => User::with('role')->superAdminCreate()]);
+        return view('auth.pages.Users.manageuser', ['users' => User::with('role')->adminCreate()]);
     }
     public function edit($id)
     {
-        //checks if user are allowed to edit accounts
+
         $this->authorize('edit-accounts');
-        //returns request user information
-        return view('auth.pages.Users.edituser', ['user' => User::findOrFail($id)]);
+        if (auth()->user()->role_id == $this->superadmin)
+            return view('auth.pages.Users.edituser', ['user' => User::findOrFail($id), 'roles' => Role::all()]);
+
+        return view('auth.pages.Users.edituser', ['user' => User::findOrFail($id), 'roles' => Role::admincreate()]);
     }
 
     public function delete($id)
@@ -33,6 +52,8 @@ class ManageuserController extends Controller
         //user register count cannot go below 2 users
         if (User::count() < 2) return redirect()->back()->withErrors(['message' => 'Your the last one dude, do not leave us']);
         //after validation, deletes it
+        if (Role::higherAuthority(auth()->user()->id, $id))
+            return redirect()->back()->withErrors(['message' => 'You dont have authority']);
         User::destroy($id);
         //redirect back
         return redirect()->route('admin.manageuser');
@@ -66,31 +87,23 @@ class ManageuserController extends Controller
         //checks if users has permission to edit account
         $this->authorize('edit-accounts');
         //set counter.
-        $counter = $email = 0;
+        $user = User::findorFail($id);
+        // dd($user->id);
+        $request->validate([
+            'name' => 'required',
+            'email' => 'required|email|unique:users,email,' . $user->id,
+            'user_name' => 'required|unique:users,user_name,' . $user->id,
+            'roles' => 'required|exists:App\Models\Role,id'
 
-
-        //checks if email is already taken
-        foreach (User::where('email', $request->email)->get() as $user) {
-            if ($user->id != $id) {
-                ++$email;
-            }
-        }
-        //checks if user_name is already taken
-        foreach (User::where('user_name', $request->user_name)->get() as $user) {
-            if ($user->id != $id) {
-                ++$counter;
-            }
-        }
-        //if counter is above 0, then users are redirected back
-        if ($counter)   return redirect()->back()->withInput()->withErrors(['message' => 'username taken']);
-        if ($email)   return redirect()->back()->withInput()->withErrors(['message' => 'Email Taken']);
-
+        ]);
         //updates user info
         if ($request->has('updateUser')) {
-            $user = User::findorFail($id);
-            $user->is_admin = $request->is_admin === null ? false : true;
-            $user->save;
-            $user->update($request->except(['_token', 'updateUser', '_method']));
+            $user->name = $request->name;
+            $user->email = $request->email;
+            $user->user_name = $request->user_name;
+            $user->role_id = intval($request->roles);
+            $user->save();
+            // $user->update($request->except(['_token', 'updateUser', '_method']));
             return redirect()->route('admin.manageuser')->with('message', 'User sucessfully updated');
         }
     }
@@ -99,6 +112,8 @@ class ManageuserController extends Controller
     public function createUser()
     {
         $this->authorize('create-accounts');
-        return view('auth.pages.Users.createuser');
+        if (auth()->user()->role_id == $this->superadmin)
+            return view('auth.pages.Users.createuser', ['roles' => Role::all()]);
+        return view('auth.pages.Users.createuser', ['roles' => Role::admincreate()]);
     }
 }
